+++ packages/@headlessui-react/src/hooks/use-did-element-move.ts
@@ -1,21 +1,21 @@
-import { useRef, type MutableRefObject } from 'react'
+import { useRef } from 'react'
 import { useIsoMorphicEffect } from './use-iso-morphic-effect'
 
-export function useDidElementMove(enabled: boolean, element: MutableRefObject<HTMLElement | null>) {
+export function useDidElementMove(enabled: boolean, element: HTMLElement | null) {
   let elementPosition = useRef({ left: 0, top: 0 })
+
   useIsoMorphicEffect(() => {
-    let el = element.current
-    if (!el) return
+    if (!element) return
 
-    let DOMRect = el.getBoundingClientRect()
+    let DOMRect = element.getBoundingClientRect()
     if (DOMRect) elementPosition.current = DOMRect
-  }, [enabled])
+  }, [enabled, element])
 
-  if (element.current == null) return false
+  if (element == null) return false
   if (!enabled) return false
-  if (element.current === document.activeElement) return false
+  if (element === document.activeElement) return false
 
-  let buttonRect = element.current.getBoundingClientRect()
+  let buttonRect = element.getBoundingClientRect()
 
   let didElementMove =
     buttonRect.top !== elementPosition.current.top ||